requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.30
python-dotenv>=1.0.0
cloudscraper>=1.2.60
requests>=2.31.0